requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.8
dm-tree==0.1.8
docutils==0.21.2
entrypoints==0.4
Faker==33.0.0
favicon==0.7.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.55.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
htbuilder==0.6.2
idna==3.6
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.0.5
keras-nightly==3.1.0.dev2024031103
kiwisolver==1.4.7
libclang==16.0.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.5.0
namex==0.0.7
narwhals==1.13.3
numpy==1.26.4
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.0
pages==0.3
pandas==2.2.3
piglet-templates==1.3.1
pillow==11.0.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.24.1
prometheus_client==0.21.0
protobuf==4.25.3
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
pymdown-extensions==10.12
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.24.0
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.40.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tb-nightly==2.17.0a20240311
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tf-nightly==2.17.0.dev20240306
tf_nightly_intel==2.17.0.dev20240306
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.2
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.2
upgrade-pip==0.1.4
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.2
upgrade-pip==0.1.4
urllib3==2.2.1
validators==0.34.0
virtualenv==20.25.1
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.2
upgrade-pip==0.1.4
urllib3==2.2.1
validators==0.34.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.2
upgrade-pip==0.1.4
tzdata==2024.2
upgrade-pip==0.1.4
upgrade-pip==0.1.4
urllib3==2.2.1
validators==0.34.0
virtualenv==20.25.1
watchdog==5.0.3
webdriver-manager==4.0.2
websocket-client==1.8.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
wrapt==1.16.0
wsproto==1.2.0
wsproto==1.2.0
zipp==3.21.0